+++ src/styles/components/Timer.styled.tsx
@@ -2,6 +2,11 @@ import styled from 'styled-components';
 
 const StyledTimer = styled.div`
   display: inline-block;
+
+  span {
+    padding-top: 1px;
+    padding-right: 4px;
+  }
 `;
 
 export default StyledTimer;